myapp/requirements.txt
awscli==1.16.125
botocore==1.12.115
dj-database-url==0.5.0
Django==2.2.3
django-cors-headers==3.0.2
django-filter==2.2.0
djangorestframework==3.7.7
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.4
pyasn1==0.4.5
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
whitenoise==4.1.3
